ci: fail on "rejected promise not handled"
Problem:
When nodejs reports "rejected promise not handled", CI does not fail.
The CI test logs must be manually inspected to notice these problems.
Requiring `await` #1638 will help, but still misses some cases.

Solution:
Fail if "rejected promise not handled" is found in the test output.
This is done in the CI script because the messages are from nodejs and
I haven't found a way to scan the nodejs output from our test runner.